feat: Add services card to admin dashboard and fix Dart errors
Flutter app improvements:
- Add "Услуги" card to admin dashboard with navigation to /admin/services
- Services screen already exists with full CRUD functionality

Bug fixes in reviews_management_screen.dart:
- Remove unused import 'package:intl/intl.dart' (line 3)
- Fix type error: change RatingWidget(rating: review.rating.toDouble())
  to RatingWidget(rating: review.rating) since rating is already int (line 124)
- RatingWidget expects int but was receiving double

Dashboard changes:
- Added services card between orders and users
- Icon: Icons.work_outline
- Navigation: /admin/services
- Grid now shows 4 cards (Orders, Services, Users, Reports)

+++ front/lib/screens/admin/reviews_management_screen.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:intl/intl.dart';
 import '../../providers/review_provider.dart';
 import '../../models/review_model.dart';
 import '../../widgets/rating_widget.dart';
@@ -121,7 +120,7 @@ class _ReviewsManagementScreenState extends State<ReviewsManagementScreen> {
                               child: Column(
                                 crossAxisAlignment: CrossAxisAlignment.start,
                                 children: [
-                                  RatingWidget(rating: review.rating.toDouble()),
+                                  RatingWidget(rating: review.rating),
                                   const SizedBox(height: 4),
                                   Text(
                                     'Клиент ID: ${review.clientId}',
